clean up indexing tests

+++ core/test/testUtils/testDir.ts
@@ -11,14 +11,16 @@ export function setUpTestDir() {
 }
 
 export function tearDownTestDir() {
-  fs.rmSync(TEST_DIR, { recursive: true });
+  if (fs.existsSync(TEST_DIR)) {
+    fs.rmSync(TEST_DIR, { recursive: true });
+  }
 }
 
 export function addToTestDir(paths: (string | string[])[]) {
   for (const p of paths) {
     const filepath = path.join(TEST_DIR, Array.isArray(p) ? p[0] : p);
     fs.mkdirSync(path.dirname(filepath), { recursive: true });
-    
+
     if (Array.isArray(p)) {
       fs.writeFileSync(filepath, p[1]);
     } else if (p.endsWith("/")) {
